+++ src/main/java/me/kyllian/netflixstatistix/controllers/LoginController.java
@@ -3,9 +3,19 @@ package me.kyllian.netflixstatistix.controllers;
 import javafx.application.Platform;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.fxml.Initializable;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
+import javafx.stage.StageStyle;
 import me.kyllian.netflixstatistix.exceptions.InputInvalidException;
 import me.kyllian.netflixstatistix.exceptions.InvalidFieldType;
 import me.kyllian.netflixstatistix.post.PasswordEncryptor;
@@ -13,21 +23,78 @@ import me.kyllian.netflixstatistix.post.PostBuilder;
 import me.kyllian.netflixstatistix.user.User;
 import me.kyllian.netflixstatistix.user.UserBuilder;
 
-public class LoginController {
+import java.io.IOException;
+import java.net.URL;
+import java.util.ResourceBundle;
 
-    private boolean loggingIn = false;
+public class LoginController implements Initializable {
 
-    @FXML
-    private Label target;
-    @FXML
-    private PasswordField passwordField;
-    @FXML
-    private TextField emailField;
+    private boolean loggingIn;
 
     @FXML
-    protected void login(ActionEvent event) {
+    private ImageView logo;
+
+    @FXML
+    private TextField email;
+
+    @FXML
+    private PasswordField password;
+
+    @FXML
+    private Label registerText;
+
+    @FXML
+    private Button loginButton;
+
+    public LoginController() {
+        this.loggingIn = false;
+    }
+
+    @Override
+    public void initialize(URL location, ResourceBundle resources) {
+        logo.setImage(new Image(getClass().getClassLoader().getResource("assets/logo.png").toExternalForm()));
+    }
+
+    public void login() {
         if (loggingIn) return;
-        target.setText("Logging in...");
         loggingIn = true;
+        loginButton.setText("Logging in...");
+        new PostBuilder().withIdentifier("login")
+                .withAttribute("email", email.getText())
+                .withAttribute("password", PasswordEncryptor.encrypt(password.getText()))
+                .post(this);
+        //TODO Actually log in
+    }
+
+    public void toRegister() {
+        try {
+            Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("view/registration.fxml"));
+            root.getStylesheets().removeAll();
+            root.getStylesheets().add(getClass().getResource("/css/registration.css").toExternalForm());
+            loginButton.getScene().setRoot(root);
+        } catch (IOException exception) {
+            exception.printStackTrace();
+        }
+    }
+
+    public void handleLogin(String response) {
+        Platform.runLater(new Runnable(){
+            @Override
+            public void run() {
+                loggingIn = false;
+                if (response.equalsIgnoreCase("INVALID_USER")) {
+                    loginButton.setText("Invalid user");
+                    return;
+                }
+                if (response.equalsIgnoreCase("INVALID_PASSWORD")) {
+                    loginButton.setText("Invalid password");
+                    return;
+                }
+                User user = new User(response);
+                loginButton.setText("Login successful");
+                //TODO: GOTO NEXT SCREEN, SAVE USER OBJECT?
+            }
+        });
+
     }
 }
