+++ src/main/resources/css/registration.css
@@ -5,3 +5,24 @@
 .label {
     -fx-text-fill: White;
 }
+
+.button {
+    -fx-background-color: #171717;
+    -fx-border-color: #717171;
+    -fx-border-radius: 5;
+    -fx-text-fill: #717171;
+}
+
+.password-field {
+    -fx-background-color: #2F2F2F;
+    -fx-background-insets: 1px;
+    -fx-prompt-text-fill: #717171;
+    -fx-text-inner-color: #717171;
+}
+
+.text-field {
+    -fx-background-color: #2F2F2F;
+    -fx-background-insets: 1px;
+    -fx-prompt-text-fill: #717171;
+    -fx-text-inner-color: #717171;
+}
