+++ src/main/resources/view/home.fxml
@@ -19,29 +19,6 @@
         <Insets top="25" right="25" bottom="10" left="25"/>
     </padding>
     <children>
-        <HBox alignment="CENTER" GridPane.rowIndex="0" GridPane.columnIndex="0">
-            <ImageView fx:id="logo" fitHeight="190.0" fitWidth="190.0"
-                       preserveRatio="true"></ImageView>
-        </HBox>
-        <VBox minWidth="350" minHeight="90" GridPane.rowIndex="1" GridPane.columnIndex="0">
-            <TextField fx:id="email" promptText="Email" minHeight="41" maxHeight="41">
 
-            </TextField>
-            <PasswordField fx:id="password" promptText="Password" minHeight="41" maxHeight="41">
-
-            </PasswordField>
-        </VBox>
-
-        <HBox alignment="CENTER" GridPane.rowIndex="2" GridPane.columnIndex="0">
-            <Button fx:id="loginButton" onMouseClicked="#login" minWidth="350" minHeight="41">
-                <text>Sign In</text>
-            </Button>
-        </HBox>
-
-        <HBox alignment="CENTER" GridPane.rowIndex="3" GridPane.columnIndex="0">
-            <Label fx:id="registerText" onMouseClicked="#toRegister">
-                <text>Nog geen account? Klik hier om te registreren</text>
-            </Label>
-        </HBox>
     </children>
 </GridPane>
