+++ src/main/resources/view/registration.fxml
@@ -1,124 +1,57 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?xml-stylesheet type="text/css" href="resources/css/registration.css"?>
 
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
-
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.text.Font?>
+<?import javafx.scene.image.ImageView?>
 <GridPane fx:controller="me.kyllian.netflixstatistix.controllers.RegistrationController"
-          xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10" styleClass="root"
-          >
+          xmlns:fx="http://javafx.com/fxml" alignment="center" styleClass="root"
+          prefHeight="500"
+          prefWidth="700"
+>
 
-    <!-- Add Header Label -->
-    <Label text="Registration Form" GridPane.columnIndex="0"
-           GridPane.rowIndex="0" GridPane.columnSpan="2"
-           GridPane.rowSpan="1" GridPane.halignment="CENTER" >
-        <font>
-            <Font name="Arial" size="24" ></Font>
-        </font>
+    <padding>
+        <Insets top="25" right="25" bottom="10" left="25"/>
+    </padding>
+
+    <ImageView fx:id="logo" fitHeight="100" fitWidth="100" GridPane.halignment="LEFT" GridPane.valignment="TOP"
+               preserveRatio="true"></ImageView>
+
+    <VBox GridPane.columnSpan="2">
+        <TextField fx:id="nameField" prefHeight="40" promptText="First and last name"
+                   GridPane.columnIndex="0" GridPane.rowIndex="1"/>
+        <TextField fx:id="birthField" prefHeight="40" promptText="Birthdate (DD-MM-YYYY)"
+                   GridPane.columnIndex="0" GridPane.rowIndex="2"/>
+        <TextField fx:id="emailField" prefHeight="40" promptText="Email"
+                       GridPane.columnIndex="0" GridPane.rowIndex="3"/>
+        <PasswordField fx:id="passwordField" prefHeight="40" promptText="Password"
+                   GridPane.columnIndex="0" GridPane.rowIndex="4"/>
+        <PasswordField fx:id="repeatPasswordField" prefHeight="40" promptText="Repeat password"
+                   GridPane.columnIndex="0" GridPane.rowIndex="5"/>
+        <TextField fx:id="streetField" prefHeight="40" promptText="Street name"
+                   GridPane.columnIndex="0" GridPane.rowIndex="6"/>
+        <TextField fx:id="numberField" prefHeight="40" promptText="Number"
+                   GridPane.columnIndex="0" GridPane.rowIndex="7"/>
+        <TextField fx:id="postalcodeField" prefHeight="40" promptText="Postal code"
+                   GridPane.columnIndex="0" GridPane.rowIndex="8"/>
+        <TextField fx:id="residenceField" prefHeight="40" promptText="Residence"
+                   GridPane.columnIndex="0" GridPane.rowIndex="9"/>
+    </VBox>
+    <Button fx:id="backButton" text="Return" onMouseClicked="#back" GridPane.columnIndex="0" GridPane.rowIndex="10"
+            prefWidth="170" prefHeight="50">
         <GridPane.margin>
-            <Insets top="20" right="0" bottom="20" left="0"></Insets>
-        </GridPane.margin>
-    </Label>
-
-
-
-    <!-- Add Name Label -->
-    <Label text="First Name : " GridPane.columnIndex="0"
-           GridPane.rowIndex="1" >
-    </Label>
-
-
-    <!-- Add Name Text Field -->
-    <TextField fx:id="nameField" prefHeight="40"
-               GridPane.columnIndex="1" GridPane.rowIndex="1"/>
-
-
-    <!-- Add Lastname Label -->
-    <Label text="Last Name : " GridPane.columnIndex="0"
-           GridPane.rowIndex="2" >
-    </Label>
-    <!-- Add Lastname Text Field -->
-    <TextField fx:id="Lastname" prefHeight="40"
-               GridPane.columnIndex="1" GridPane.rowIndex="2"/>
-
-
-
-    <!-- Add Password Label -->
-    <Label text="Password : " GridPane.columnIndex="0"
-           GridPane.rowIndex="3" >
-    </Label>
-    <!-- Add Password Field -->
-    <PasswordField fx:id="passwordField" prefHeight="40"
-                   GridPane.columnIndex="1" GridPane.rowIndex="3"/>
-
-    <!-- Add Email Label -->
-    <Label text="Email : " GridPane.columnIndex="0"
-           GridPane.rowIndex="4" >
-    </Label>
-    <!-- Add Email Field -->
-    <TextField fx:id="Email" prefHeight="40"
-                   GridPane.columnIndex="1" GridPane.rowIndex="4"/>
-
-    <!-- Add Street Label -->
-    <Label text="Street : " GridPane.columnIndex="0"
-           GridPane.rowIndex="5" >
-    </Label>
-    <!-- Add Street Field -->
-    <TextField fx:id="Street" prefHeight="40"
-                   GridPane.columnIndex="1" GridPane.rowIndex="5"/>
-
-    <!-- Add Housenumber Label -->
-    <Label text="House Number : " GridPane.columnIndex="0"
-           GridPane.rowIndex="6" >
-    </Label>
-    <!-- Add Housenumber Field -->
-    <TextField fx:id="Housenumber" prefHeight="40"
-                   GridPane.columnIndex="1" GridPane.rowIndex="6"/>
-
-    <!-- Add Birthdate Label -->
-    <Label text="Birthdate : " GridPane.columnIndex="0"
-           GridPane.rowIndex="7" >
-    </Label>
-    <!-- Add Housenumber Field -->
-    <TextField fx:id="Birthdate" prefHeight="40"
-                   GridPane.columnIndex="1" GridPane.rowIndex="7"/>
-
-
-
-
-
-
-
-    <!-- Add Submit Button -->
-    <Button fx:id="submitButton" text="Submit"
-            prefWidth="100" prefHeight="40" defaultButton="true"
-            GridPane.columnIndex="0" GridPane.rowIndex="8"
-            GridPane.columnSpan="2" GridPane.rowSpan="1"
-            GridPane.halignment="CENTER">
-            <!-- moet nog een onAction komen hier -->
-        <GridPane.margin>
-            <Insets top="20" right="0" bottom="20" left="0"></Insets>
+            <Insets top="10"/>
         </GridPane.margin>
     </Button>
-
-
-    <Label text="Registration Form" GridPane.columnIndex="0"
-           GridPane.rowIndex="0" GridPane.columnSpan="2"
-           GridPane.rowSpan="1" GridPane.halignment="CENTER" >
-        <font>
-            <Font name="Arial" size="24" ></Font>
-        </font>
+    <Button fx:id="submitButton" text="Create account"
+            prefWidth="170" prefHeight="50"
+            GridPane.columnIndex="1" GridPane.rowIndex="10"
+            GridPane.halignment="CENTER">
         <GridPane.margin>
-            <Insets top="20" right="0" bottom="20" left="0"></Insets>
+            <Insets top="10"/>
         </GridPane.margin>
-    </Label>
-
-
+    </Button>
 
 
 </GridPane>
